requirements.txt
apns2==0.4.1
asn1crypto==0.24.0
boto==2.49.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
dj-database-url==0.5.0
Django==2.1.2
django-appconf==1.0.2
django-countries==5.3.2
django-countries-plus==1.2.1
django-model-utils==3.1.2
django-mysql==2.4.1
django-notifications-hq==1.5.0
django-paypal==0.5.0
django-progressbarupload==0.1.7
django-push-notifications==1.6.0
django-storages==1.7.1
djangorestframework==3.9.0
easy-thumbnails==2.5
gunicorn==19.9.0
h2==2.6.2
hpack==3.0.0
http-ece==1.0.5
hyper==0.7.0
hyperframe==3.2.0
idna==2.7
jsonfield==2.0.2
notifications==0.3.2
olefile==0.46
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py-vapid==1.4.0
pycparser==2.19
PyJWT==1.6.4
pytz==2018.5
pywebpush==1.7.0
requests==2.19.1
six==1.11.0
urllib3==1.23
whitenoise==3.3.0
